refactor(web): centralize studio LoRA selection + extract picker/save panels (sc-4196)
ImageStudio and VideoStudio duplicated the entire LoRA selection machinery —
compatibleLoras/compatibleLoraKey/selectedLoras/selectedLoraValidationResult/
loraEmptyMessage, the prune-on-model-change and auto-open-advanced effects, and
toggleLora/effectiveLoraWeight/setLoraWeight (the 2-user-LoRA cap rule lived in
three places) — plus near-identical LoRA-picker and Save-as-Preset JSX. ~300
duplicated lines patched in tandem several times, divergence risk on every change.

- Move the LoRA selection state (selectedLoraIds/loraWeights/showIncompatibleLoras,
  seeded from the persisted snapshot) + derived values + actions + the two effects
  into useGenerationStudio (which both studios already call). It returns the bundle
  under the same names the studios used, so persistence/launch/generate code is
  unchanged; advancedOpen/setAdvancedOpen are passed in for the auto-open effect.
- Extract <LoraPickerSection> (the compatible-LoRA checklist + weight sliders +
  empty state) and <SavePresetPanel> (name field, scope segment, message) to
  generationStudio.jsx, alongside the existing PresetGuidanceStrip/PresetValidation-
  Warnings. handleSaveAsPreset stays per-studio (different payloads); the video
  studio's mode-gated save disable + tooltip ride along as optional props.

Behavior preserved. Full web suite 377 passed (ImageStudio/VideoStudio + the
170-test main suite exercise the picker + preset save through the shared code);
vite build clean. Studios shrink ~380 dup lines for +259 shared.
